+++ sso/models.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import AbstractUser
 from django.db import models
 from django.utils.functional import cached_property
 
-from directory_components.helpers import CompanyParser
+from great_components.helpers import CompanyParser
 
 from sso import helpers